+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Verification is no longer performed on macro expansion. Instead,
+  during building the segments, any segment that doesn't pass a
+  defined verification is not added. Furthermore, such a segment's (or
+  augment's) setup function will return early as well.
+
 ## [v0.6.0]
 
 ### Changed